chore(packages): some file structure improvement to the graphql client (#692)
