+++ services/vault/src/clients/graphql/index.ts
@@ -1,0 +1,1 @@
+export { graphqlClient } from "./client";
