+++ services/vault/src/registry/index.ts
@@ -1,0 +1,5 @@
+export {
+  APPLICATION_REGISTRY,
+  getApplicationMetadata,
+  type ApplicationMetadata,
+} from "./applications";
