+++ services/vault/src/services/applications/index.ts
@@ -1,0 +1,1 @@
+export { fetchApplications } from "./fetchApplications";
